Moving functions to prevent circular import
